+++ api/v1alpha1/podhttpchaos_types.go
@@ -26,13 +26,21 @@ type PodHttpChaosSpec struct {
 
 // PodHttpChaosStatus defines the actual state of PodHttpChaos.
 type PodHttpChaosStatus struct {
+	//+kubebuilder:deprecatedversion:warning="Pid is deprecated, please use Uid instead"
 	// Pid represents a running tproxy process id.
+	// Deprecated: Pid is deprecated, please use Uid instead
 	// +optional
 	Pid int64 `json:"pid,omitempty"`
 
+	//+kubebuilder:deprecatedversion:warning="StartTime is deprecated"
 	// StartTime represents the start time of a tproxy process.
+	// Deprecated: StartTime is deprecated
 	// +optional
 	StartTime int64 `json:"startTime,omitempty"`
+
+	// Uid represents a unique running chaos-proxy process
+	// +optional
+	Uid string `json:"uid,omitempty"`
 
 	// +optional
 	FailedMessage string `json:"failedMessage,omitempty"`