+++ api/v1alpha1/podiochaos_types.go
@@ -227,13 +227,21 @@ type PodIOChaos struct {
 
 type PodIOChaosStatus struct {
 
+	//+kubebuilder:deprecatedversion:warning="Pid is deprecated, please use Uid instead"
 	// Pid represents a running toda process id
+	// Deprecated: Pid is deprecated, please use Uid instead
 	// +optional
 	Pid int64 `json:"pid,omitempty"`
 
+	//+kubebuilder:deprecatedversion:warning="StartTime is deprecated"
 	// StartTime represents the start time of a toda process
+	// Deprecated: StartTime is deprecated
 	// +optional
 	StartTime int64 `json:"startTime,omitempty"`
+
+	// Uid represents a unique running toda process
+	// +optional
+	Uid string `json:"uid,omitempty"`
 
 	// +optional
 	FailedMessage string `json:"failedMessage,omitempty"`