+++ pkg/chaosdaemon/iochaos_server.go
@@ -39,12 +39,6 @@ func (s *DaemonServer) ApplyIOChaos(ctx context.Context, in *pb.ApplyIOChaosRequ
 	log := s.getLoggerFromContext(ctx)
 	log.Info("applying io chaos", "Request", in)
 
-	if in.InstanceUid == "" {
-		if uid, ok := s.backgroundProcessManager.GetUID(bpm.ProcessPair{Pid: int(in.Instance), CreateTime: in.StartTime}); ok {
-			in.InstanceUid = uid
-		}
-	}
-
 	if in.InstanceUid != "" {
 		err := s.killIOChaos(ctx, in.InstanceUid)
 		if err != nil {
@@ -60,10 +54,7 @@ func (s *DaemonServer) ApplyIOChaos(ctx context.Context, in *pb.ApplyIOChaosRequ
 
 	log.Info("the length of actions", "length", len(actions))
 	if len(actions) == 0 {
-		return &pb.ApplyIOChaosResponse{
-			Instance:  0,
-			StartTime: 0,
-		}, nil
+		return &pb.ApplyIOChaosResponse{}, nil
 	}
 
 	pid, err := s.crClient.GetPidFromContainerID(ctx, in.ContainerId)
@@ -114,8 +105,6 @@ func (s *DaemonServer) ApplyIOChaos(ctx context.Context, in *pb.ApplyIOChaosRequ
 	}
 
 	return &pb.ApplyIOChaosResponse{
-		Instance:    int64(proc.Pair.Pid),
-		StartTime:   proc.Pair.CreateTime,
 		InstanceUid: proc.Uid,
 	}, nil
 }